Implement delete function

+++ app/Http/Controllers/AnswersController.php
@@ -63,6 +63,9 @@ class AnswersController extends Controller
      */
     public function destroy(Question $question, Answer $answer)
     {
+        
+        $this->authorize('delete', $answer);
+        
         $answer->delete();
 
         return back()->with('success', 'Answer deleted successfully.');
